backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==23.0.0
whitenoise==6.7.0
django-filter==24.3
